move init() method from ConfigCenterConfig
