+++ dubbo-config/dubbo-config-api/src/main/java/org/apache/dubbo/config/ConfigCenterConfig.java
@@ -19,18 +19,11 @@ package org.apache.dubbo.config;
 import org.apache.dubbo.common.Constants;
 import org.apache.dubbo.common.URL;
 import org.apache.dubbo.common.config.Environment;
-import org.apache.dubbo.common.extension.ExtensionLoader;
 import org.apache.dubbo.common.utils.StringUtils;
 import org.apache.dubbo.common.utils.UrlUtils;
 import org.apache.dubbo.config.support.Parameter;
-import org.apache.dubbo.configcenter.DynamicConfiguration;
-import org.apache.dubbo.configcenter.DynamicConfigurationFactory;
 
-import java.io.IOException;
-import java.io.StringReader;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.Properties;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
@@ -60,40 +53,7 @@ public class ConfigCenterConfig extends AbstractConfig {
     public ConfigCenterConfig() {
     }
 
-    public void init(ApplicationConfig applicationConfig) {
-        if (isValid()) {
-            if (!inited.compareAndSet(false, true)) {
-                return;
-            }
-            DynamicConfiguration dynamicConfiguration = startDynamicConfiguration(toConfigUrl());
-            String configContent = dynamicConfiguration.getConfig(configFile, group);
-
-            String appGroup = applicationConfig != null ? applicationConfig.getName() : null;
-            String appConfigContent = null;
-            if (StringUtils.isNotEmpty(appGroup)) {
-                appConfigContent = dynamicConfiguration.getConfig
-                        (StringUtils.isNotEmpty(appConfigFile) ? appConfigFile : configFile,
-                         appGroup
-                        );
-            }
-            try {
-                Environment.getInstance().setConfigCenterFirst(highestPriority);
-                Environment.getInstance().updateExternalConfigurationMap(parseProperties(configContent));
-                Environment.getInstance().updateAppExternalConfigurationMap(parseProperties(appConfigContent));
-            } catch (IOException e) {
-                throw new IllegalStateException("Failed to parse configurations from Config Center.", e);
-            }
-        }
-    }
-
-    private DynamicConfiguration startDynamicConfiguration(URL url) {
-        DynamicConfigurationFactory dynamicConfigurationFactory = ExtensionLoader.getExtensionLoader(DynamicConfigurationFactory.class).getExtension(url.getProtocol());
-        DynamicConfiguration configuration = dynamicConfigurationFactory.getDynamicConfiguration(url);
-        Environment.getInstance().setDynamicConfiguration(configuration);
-        return configuration;
-    }
-
-    private URL toConfigUrl() {
+    public URL toUrl() {
         Map<String, String> map = this.getMetaData();
         if (StringUtils.isEmpty(address)) {
             address = Constants.ANYHOST_VALUE;
@@ -106,18 +66,8 @@ public class ConfigCenterConfig extends AbstractConfig {
         return UrlUtils.parseURL(address, map);
     }
 
-    protected Map<String, String> parseProperties(String content) throws IOException {
-        Map<String, String> map = new HashMap<>();
-        if (StringUtils.isEmpty(content)) {
-            logger.warn("You specified the config centre, but there's not even one single config item in it.");
-        } else {
-            Properties properties = new Properties();
-            properties.load(new StringReader(content));
-            properties.stringPropertyNames().forEach(
-                    k -> map.put(k, properties.getProperty(k))
-            );
-        }
-        return map;
+    public boolean checkOrUpdateInited() {
+        return inited.compareAndSet(false, true);
     }
 
     public void setExternalConfig(Map<String, String> externalConfiguration) {
@@ -182,7 +132,7 @@ public class ConfigCenterConfig extends AbstractConfig {
     }
 
     @Parameter(key = Constants.CONFIG_ENABLE_KEY, useKeyAsProperty = false)
-    public Boolean getHighestPriority() {
+    public Boolean isHighestPriority() {
         return highestPriority;
     }
